+++ src/egeo-demo/st-bubble-demo/st-bubble-demo.component.ts
@@ -1,4 +1,4 @@
-/**
+/*
  * © 2017 Stratio Big Data Inc., Sucursal en España.
  *
  * This software is licensed under the Apache License, Version 2.0.
@@ -8,7 +8,19 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-.content {
-   height: 0;
-}
+import { Component } from '@angular/core';
 
+@Component({
+   selector: 'st-bubble-demo',
+   templateUrl: './st-bubble-demo.component.html',
+   styleUrls: ['./st-bubble-demo.component.scss']
+})
+export class StBubbleDemoComponent {
+
+   public text: string = 'What needed to deploy Crossdata?';
+   public hidden: boolean = true;
+
+   onClick(): void {
+      this.hidden = !this.hidden;
+   }
+}