+++ src/lib/public_api.ts
@@ -50,6 +50,9 @@ export { StAlertsService } from './st-alerts/st-alerts.service';
 // Breadcrumb
 export { StBreadcrumbsModule } from './st-breadcrumbs/st-breadcrumbs.module';
 
+// Bubble
+export { StBubbleModule } from './st-bubble/st-bubble.module';
+
 // Checkbox
 export { StCheckboxModule } from './st-checkbox/st-checkbox.module';