+++ src/lib/st-bubble/st-bubble.module.ts
@@ -1,4 +1,4 @@
-/**
+/*
  * © 2017 Stratio Big Data Inc., Sucursal en España.
  *
  * This software is licensed under the Apache License, Version 2.0.
@@ -8,7 +8,15 @@
  *
  * SPDX-License-Identifier: Apache-2.0.
  */
-.content {
-   height: 0;
-}
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
 
+import { StBubbleComponent } from './st-bubble.component';
+import { StPopModule } from '../st-pop/st-pop.module';
+
+@NgModule({
+   imports: [CommonModule, StPopModule],
+   declarations: [StBubbleComponent],
+   exports: [StBubbleComponent]
+})
+export class StBubbleModule {}